fix: base on codgen for ShoppingCart in eventing example
