+++ samples/java-eventing-shopping-cart/src/main/proto/cart/shopping_cart_domain.proto
@@ -27,7 +27,7 @@ import "akkaserverless/annotations.proto";
 // end::events-and-state[]
 
 option (akkaserverless.file).event_sourced_entity = {
-  name: "ShoppingCart"
+  name: "ShoppingCartEntity"
   entity_type: "eventsourced-shopping-cart"
   state: "CartState"
   events: ["ItemAdded", "ItemRemoved", "CheckedOut"]
