+++ samples/java-eventing-shopping-cart/src/main/proto/product/product_popularity_domain.proto
@@ -23,7 +23,7 @@ option java_outer_classname = "ProductPopularityDomain";
 import "akkaserverless/annotations.proto";
 
 option (akkaserverless.file).value_entity = {
-  name: "ProductPopularity"
+  name: "ProductPopularityEntity"
   entity_type: "product-popularity"
   state: "Popularity"
 };
